% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cor.mtest.R
\name{cor.mtest}
\alias{cor.mtest}
\title{cor.mtest}
\usage{
cor.mtest(mat, ...)
}
\arguments{
\item{mat}{matrix to be tested, col.names will be correlated}

\item{...}{further arguments to pass to the native R cor.test function}
}
\value{

}
\description{
To compute the matrix of p-value of the correlation test
Credit: http://www.sthda.com/english/wiki/visualize-correlation-matrix-using-correlogram
}
\examples{
cor.mtest(mtcars)
}
