% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.import.R
\name{fread.delim}
\alias{fread.delim}
\title{Fast read data
Fast read data frame
The functions are developed to fit personal preference. Feel free to modify it.}
\usage{
fread.delim(path_to_file, delim = "\\t", header = T, cores = 29)
}
\arguments{
\item{path_to_file:}{the path to the dataset}

\item{delim:}{delimiter, default (tab-delimited)}

\item{header:}{default TRUE}

\item{cores:}{number of cores when reading the dataset. default: 29.  Why 29? Because I love the prime numbers.}
}
\value{

}
\description{
Fast read data
Fast read data frame
The functions are developed to fit personal preference. Feel free to modify it.
}
\examples{
fread.delim(path_to_file = data.frame(Gene.Symbol = c("a", "b", "c"),
Cell1 = c(1, 1, 2), Cell2 = c(2, 2, 4), stringsAsFactors  = F, check.names = F))
}
