% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.import.R
\name{fread.matrix}
\alias{fread.matrix}
\title{Title
fast read a dataset into a matrix}
\usage{
fread.matrix(path_to_file, delim = "auto", header = "auto", cores = 29)
}
\arguments{
\item{path_to_file:}{the path to the dataset}

\item{delim:}{delimiter, default (tab-delimited)}

\item{header:}{default TRUE}

\item{cores:}{number of cores when reading the dataset. default: 29.  Why 29? Because I love the prime numbers.}
}
\value{

}
\description{
Title
fast read a dataset into a matrix
}
\examples{
fread.matrix('https://raw.githubusercontent.com/jiayiliujiayi/timekiller/master/testdata/dataset_matrix.txt')
}
