% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.import.R
\name{fread.matrix2}
\alias{fread.matrix2}
\title{Fast read matrix with unique rownames
fast read a dataset into a matrix}
\usage{
fread.matrix2(
  path_to_file,
  method = sum,
  delim = "auto",
  header = "auto",
  cores = 29,
  blank = "novel"
)
}
\arguments{
\item{path_to_file}{the path to the dataset}

\item{method}{defalut = "sum"}

\item{delim}{delimiter, default (tab-delimited)}

\item{header}{default TRUE}

\item{cores}{number of cores when reading the dataset. default: 29.  Why 29? Because I love the prime numbers.}

\item{blank}{rename the blank rownames as}
}
\value{

}
\description{
Fast read matrix with unique rownames
fast read a dataset into a matrix
}
