% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.inspection.R
\name{getCharCols}
\alias{getCharCols}
\title{getCharCols}
\usage{
getCharCols(path_to_file)
}
\arguments{
\item{path_to_file}{}
}
\value{

}
\description{
Extract the colnames of the "character" columns
}
\examples{
getCharColsifCharCols('https://raw.githubusercontent.com/jiayiliujiayi/timekiller/master/testdata/dataset_df.txt')
}
