% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.inspection.R
\name{getNACols}
\alias{getNACols}
\title{getNACols
Extract the colnames of the NA-containing columns in your dataset.}
\usage{
getNACols(path_to_file)
}
\arguments{
\item{path_to_file}{the path to the dataset, or the name of the element in the working environment}
}
\value{

}
\description{
getNACols
Extract the colnames of the NA-containing columns in your dataset.
}
\examples{
getNACols('https://raw.githubusercontent.com/jiayiliujiayi/timekiller/master/testdata/dataset_df.txt')
}
