% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.getNature.R
\name{getNature}
\alias{getNature}
\title{getNature}
\usage{
getNature(category = "news")
}
\arguments{
\item{category}{which category of the news would you like to check?
Options: ("news", "research" or "opinion"):
"news" (default), the latest Nature news.
"research": the latest research articles and review articles from the Nature journal.
"opinion": the latest opinion articles.}
}
\value{
PeoplesDailyNews
}
\description{
Get the news from the Nature
}
\examples{
getNature()
}
\keyword{news}
