% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.getPeoplesDaily.R
\name{getPeoplesDaily}
\alias{getPeoplesDaily}
\title{get the news from the People's Daily}
\usage{
getPeoplesDaily(category = "headline")
}
\arguments{
\item{category}{which category of the news would you like to check?
Defaluts to "headline", could be "headline", "tophit" or "latest"}
}
\value{
PeoplesDailyNews
}
\description{
get the news from the People's Daily
}
\examples{
getPeoplesDaily()
}
\keyword{news}
