% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.getScience.R
\name{getScience}
\alias{getScience}
\title{getScience}
\usage{
getScience(category = "latest")
}
\arguments{
\item{category}{default = "latest".  options: "latest", "insider"}
}
\value{

}
\description{
Get the news from the Science magazine.
}
\examples{
getScience()
}
