% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.inspection.R
\name{ifCharCols}
\alias{ifCharCols}
\title{ifCharCols}
\usage{
ifCharCols(path_to_file)
}
\arguments{
\item{path_to_file}{}
}
\value{

}
\description{
Check if there are character columns in your dataset.
}
\examples{
ifCharCols('https://raw.githubusercontent.com/jiayiliujiayi/timekiller/master/testdata/dataset_df.txt')
}
