% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.inspection.R
\name{ifNACols}
\alias{ifNACols}
\title{ifNACols}
\usage{
ifNACols(path_to_file)
}
\arguments{
\item{path_to_file}{the path to the dataset, or the name of the element in the working environment}
}
\value{

}
\description{
Check if there are NA-containing columns in your dataset.
}
\examples{
ifNACols('https://raw.githubusercontent.com/jiayiliujiayi/timekiller/master/testdata/dataset_df.txt')
}
