% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.myLuckyNumber.R
\name{myLuckyNumber}
\alias{myLuckyNumber}
\title{myLuckyNumber}
\usage{
myLuckyNumber(min = 1, max = 100)
}
\arguments{
\item{min}{lower limit of the number, default 1}

\item{max}{upper limit of the number, defalut 2}
}
\value{

}
\description{
Ask the timekiller what's your lucknumber for today.  Remeber not to ask it two times in the same day.
}
