% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.play24points.R
\name{play24points}
\alias{play24points}
\title{play24points}
\usage{
play24points(max.value = 10, min.value = 1)
}
\arguments{
\item{max.value}{max value in the pocker cards, default 1}

\item{min.value}{min value in the pocker cards, default 5}
}
\value{

}
\description{
Randomly pick 4 cards.  You can define the min and max values of the cards to be picked!
}
\examples{
play24points()

}
