% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.scrna10x.R
\name{raw10xtoDF}
\alias{raw10xtoDF}
\title{Title}
\usage{
raw10xtoDF(path_to_file, id = "Gene Symbol", cores = 19,
  delim = "auto", header = T)
}
\arguments{
\item{path_to_file}{the path to the dataset}

\item{id}{the grouping id in the dataset, default = "Gene Symbol".}

\item{cores}{number of cores when reading the dataset. default: 19.}

\item{delim}{delimiter, default (tab-delimited)}

\item{header}{default = T}
}
\description{
Title
}
\examples{
raw10xtoDF

}
\keyword{data}
