% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.tobeornottobe.R
\name{tobeornottobe}
\alias{tobeornottobe}
\title{tobeornottobe}
\usage{
tobeornottobe(input = "Should I study today?")
}
\arguments{
\item{input}{ask a yes or no question or say some random words to the timekiller,
e.g.: simply "Should I have lunch today?" or just "asdklfja2".}
}
\value{

}
\description{
"To be, or not to be: that is the question. \cr
Whether it’s nobler in the mind to suffer \cr
The slings and arrows of outrageous fortune, \cr
Or to take arms against a sea of troubles, \cr
And by opposing end them?" \cr
\cr
--Hamlet 3.1
}
\examples{
tobeornottobe("Should I have lunch today?")
}
\keyword{fun}
