% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio.write.gmt.R
\name{write.gmt}
\alias{write.gmt}
\title{write.gmt
write out self-curating gmt files.}
\usage{
write.gmt(geneSet = gslist, gmt_file = "gs.gmt")
}
\arguments{
\item{geneSet}{a list element, with names as geneset names.}

\item{gmt_file}{output /gmt file.  default: "gs.gmt"}
}
\value{

}
\description{
write.gmt
write out self-curating gmt files.
}
