% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.getPeoplesDaily.R
\name{getPeoplesDaily}
\alias{getPeoplesDaily}
\title{getPeoplesDaily}
\usage{
getPeoplesDaily(category = "headline")
}
\arguments{
\item{category}{which category of the news would you like to check?
Options: "headline", "tophit" or "latest" (default: "headline")}
}
\value{
PeoplesDailyNews
}
\description{
Get the news from the People's Daily
}
\examples{
getPeoplesDaily()
}
\keyword{news}
