% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.inspection.R
\name{ifNaCols}
\alias{ifNaCols}
\title{ifNaCols}
\usage{
ifNaCols(path_to_file)
}
\arguments{
\item{path_to_file}{the path to the dataset, or the name of the element in the working environment}
}
\value{

}
\description{
ifNaCols
}
\examples{
ifNaCols('https://raw.githubusercontent.com/jiayiliujiayi/timekiller/master/testdata/dataset_df.txt')
}
